document_analyser_mcp/requirements.txt
fastmcp>=0.1.0
requests>=2.31.0
keybert>=0.8.0
textstat>=0.7.3
spacy>=3.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
